+++ src/Event/BeforeApplicationInitializationEvent.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Castor\Event;
+
+use Castor\Console\Application;
+
+class BeforeApplicationInitializationEvent
+{
+    public function __construct(
+        public readonly Application $application,
+    ) {
+    }
+}