+++ src/Remote/Exception/ComposerError.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Castor\Remote\Exception;
+
+class ComposerError extends \RuntimeException
+{
+}