+++ tests/Helper/OutputCleaner.php
@@ -16,6 +16,7 @@ final class OutputCleaner
         $string = preg_replace('{In ContextRegistry.php line \d+:}m', 'In ContextRegistry.php line XXXX:', $string);
         $string = preg_replace('{you are using v\d+.\d+.\d+.}m', 'you are using vX.Y.Z.', $string);
         $string = preg_replace('{^\d\d:\d\d:\d\d }m', 'hh:mm:ss ', $string);
+        $string = preg_replace("{\d{2}:\d{2}:\d{2} WARNING   \[castor\] Remote imports are disabled, skipping import of \".*\".\n}m", '', $string);
 
         // Avoid spacing issues
         $string = ltrim($string, "\n"); // Trim output start to avoid empty lines